+++ app/src/main/kotlin/org/dbtools/sample/roomsqlite/App.kt
@@ -1,5 +1,0 @@
-package org.dbtools.sample.roomsqlite
-
-import android.app.Application
-
-class App : Application()