+++ dbtools-room/src/commonMain/kotlin/org/dbtools/room/DatabaseViewQuery.kt
@@ -1,0 +1,3 @@
+package org.dbtools.room
+
+class DatabaseViewQuery(val viewName: String, val viewQuery: String)