+++ dbtools-room/src/commonMain/kotlin/org/dbtools/room/log/DbToolsRoomCrashLogException.kt
@@ -1,0 +1,3 @@
+package org.dbtools.room.log
+
+class DbToolsRoomCrashLogException : RuntimeException()