+++ library/src/main/java/org/dbtools/android/room/DatabaseViewQuery.kt
@@ -1,3 +1,0 @@
-package org.dbtools.android.room
-
-class DatabaseViewQuery(val viewName: String, val viewQuery: String)