+++ library/src/main/java/org/dbtools/android/room/log/DbToolsRoomCrashLogException.kt
@@ -1,3 +1,0 @@
-package org.dbtools.android.room.log
-
-class DbToolsRoomCrashLogException : RuntimeException()